fix(p1): correct _mm_crc32_u8 arg type u8 (was u32)
GH build 33799278446 failed E0308 expected u8 found u32 at sink.rs:49
Fix: _mm_crc32_u8(crc, bytes[i]) not as u32

No other changes; keep P0 x86-64-v3 + P1 CRC32C fast path

+++ crates/nf-testkit/src/sink.rs
@@ -46,7 +46,7 @@ fn fast_hash_bytes(mut h: u64, bytes: &[u8]) -> u64 {
     }
     if i < len {
         unsafe {
-            crc = std::arch::x86_64::_mm_crc32_u8(crc, bytes[i] as u32);
+            crc = std::arch::x86_64::_mm_crc32_u8(crc, bytes[i]);
         }
     }
     let crc = crc ^ 0xffffffffu32;
